Create proxy User model; delete unnecessary models; edit manager clases

+++ project/sound_store/SoundManager.py
@@ -1,11 +1,12 @@
-from .models import Users, Sounds
+from .models import UserData, Sounds
 from .src.DataManager import DataManager
 from .BasicManager import BasicManager, STORE_PATH
 
 
 class SoundManager(BasicManager):
     def get(self, user_id):
-        user = Users.get_user_object(user_id)
+        result = None
+        user = UserData.get_user_object(user_id)
         if user is not None:
             manager = DataManager(user_id, STORE_PATH)
             manager.set_user_id(user_id)
@@ -13,7 +14,8 @@ class SoundManager(BasicManager):
         return result
 
     def create(self, user_id, sound_name, file):
-        user = Users.get_user_object(user_id)
+        result = None
+        user = UserData.get_user_object(user_id)
         print(user)
         if user is not None:
             manager = DataManager(user_id, STORE_PATH)
@@ -26,14 +28,17 @@ class SoundManager(BasicManager):
         return result
 
     def delete(self, user_id, sound_name):
-        user = Users.get_user_object(user_id)
+        result = None
+        user = UserData.get_user_object(user_id)
         if user is not None:
             manager = DataManager(user_id, STORE_PATH)
             manager.set_user_id(user_id)
-            manager.delete_user_file(sound_name)  # FIXME if sound already deleted
-            result = Sounds.sound_data_by_name(user_id, sound_name)
-            sound = Sounds.get_sound_object(result['id'])
-            sound.delete()
+            is_deleted = manager.delete_user_file(sound_name)
+            if is_deleted is True:
+                result = Sounds.sound_data_by_name(user_id, sound_name)
+                sound = Sounds.get_sound_object(result['id'])
+                if sound is not None:
+                    sound.delete()
         return result
 
     def update(self, *args):
@@ -41,14 +46,13 @@ class SoundManager(BasicManager):
 
     def load(self, user_id, sound_name):
         result = None
-        user = Users.get_user_object(user_id)
+        user = UserData.get_user_object(user_id)
         if user is not None:
             manager = DataManager(user_id, STORE_PATH)
             manager.set_user_id(user_id)
             file_name = manager.get_full_file_path(sound_name)
             if file_name:
                 file_object = open(file_name, 'rb')
-                data = Sounds.sound_data_by_name(7, "oo")
-                print(data)
+                data = Sounds.sound_data_by_name(user_id, file_name)
                 result = file_object
         return result
